requirements.txt
requests
pandas
pymongo
mysql-connector-python
